+++ lib/ruby_speech/nlsml.rb
@@ -2,8 +2,7 @@ module RubySpeech
   module NLSML
     extend ActiveSupport::Autoload
 
-    NLSML_NAMESPACE   = 'http://www.w3c.org/2000/11/nlsml'
-    XFORMS_NAMESPACE  = 'http://www.w3.org/2000/xforms'
+    NLSML_NAMESPACE = 'http://www.ietf.org/xml/ns/mrcpv2'
 
     eager_autoload do
       autoload :Builder