+++ lib/ruby_speech/nlsml/builder.rb
@@ -4,7 +4,7 @@ module RubySpeech
       attr_reader :document
 
       def initialize(options = {}, &block)
-        options = {'xmlns' => NLSML_NAMESPACE, 'xmlns:xf' => XFORMS_NAMESPACE}.merge(options)
+        options = {'xmlns' => NLSML_NAMESPACE}.merge(options)
         @document = Nokogiri::XML::Builder.new do |builder|
           builder.result options do |r|
             apply_block r, &block
@@ -14,17 +14,9 @@ module RubySpeech
 
       def interpretation(*args, &block)
         if args.last.respond_to?(:has_key?) && args.last.has_key?(:confidence)
-          args.last[:confidence] = (args.last[:confidence] * 100).to_i
+          args.last[:confidence] = args.last[:confidence].to_f
         end
         @result.send :interpretation, *args, &block
-      end
-
-      def model(*args, &block)
-        xf_namespaced_element :model, *args, &block
-      end
-
-      def instance(*args, &block)
-        xf_namespaced_element :instance, *args, &block
       end
 
       def method_missing(method_name, *args, &block)
@@ -36,11 +28,6 @@ module RubySpeech
       def apply_block(result, &block)
         @result = result
         instance_eval &block
-      end
-
-      def xf_namespaced_element(element_name, *args, &block)
-        namespace = @result.send :[], 'xf'
-        namespace.send element_name, &block
       end
     end
   end